feat: Update Summoner class to include all necessary attributes
This commit updates the Summoner class in the riot/models/summoner.py file to include all the necessary attributes for representing summoner information. The __init__ method now accepts the id, accountId, puuid, profileIconId, revisionDate, and summonerLevel parameters and assigns them to the corresponding instance variables. The __repr__ and __str__ methods have also been implemented to provide a string representation of the Summoner object.